feat: resetPasswordメソッドの作成

+++ client/src/libs/auth/useUser.ts
@@ -5,6 +5,7 @@ import { emitError } from '../../redux/modules/dialog';
 import initFirebase from './initFirebase';
 import { alertError } from './alertError';
 import { remove, update } from '../../redux/modules/users';
+import { removeUserCookie } from './userCookies';
 
 // TODO: 型修正
 export const useUser = (): any => {
@@ -14,11 +15,10 @@ export const useUser = (): any => {
   const router = useRouter();
   const dispatch = useDispatch();
 
-  const user = firebase.auth().currentUser;
-
-  if (!user) return null;
-
   const updateUsername = async (username: string) => {
+    const user = firebase.auth().currentUser;
+
+    if (!user) return;
     try {
       await user.updateProfile({ displayName: username });
       dispatch(update({ id: user.uid, username }));
@@ -27,14 +27,30 @@ export const useUser = (): any => {
     }
   };
 
-  const deleteUser = async () => {
+  const resetPassword = async (email: string) => {
+    await firebase.auth().sendPasswordResetEmail(email);
+  };
+
+  const deleteUser = async (value: { email: string; password: string }) => {
+    const user = firebase.auth().currentUser;
+
+    if (!user) return;
+    const { email, password } = value;
     try {
+      const credential = firebase.auth.EmailAuthProvider.credential(
+        email,
+        password,
+      );
+      if (!credential) return;
+      await user.reauthenticateWithCredential(credential);
       await user.delete();
-      dispatch(remove({ id: user.uid }));
       router.push('/');
+      dispatch(remove({ id: user.uid }));
+      removeUserCookie();
     } catch (error) {
       dispatch(emitError(alertError(error)));
     }
   };
-  return { updateUsername, deleteUser };
+
+  return { updateUsername, resetPassword, deleteUser };
 };
